chore: Move json to shared
We still keep the frozen json in Replicache because it is currently only
used there.
